Se realiza la implementaciÂón del modulo de auditorias para la recuperación de contraseñas

+++ src/main/java/com/uniminuto/clinica/repository/MedicoRepository.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
  * @author lmora
  */
 @Repository
-public interface MedicoRepository extends JpaRepository<Medico, Integer> {
+public interface MedicoRepository extends JpaRepository<Medico, Long> {
     List<Medico> findByEspecializacion(Especializacion e);
 
     Optional<Medico> findByDocumento(String numeroDocumento);
